chore(management/component): refactor and optimization
