+++ src/components/Management/subcomps/InvestmentFormComponent.jsx
@@ -1,74 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import Inputs from "./subcomps_forms/imputs";
-import { useCookies } from "react-cookie";
-import { useDispatch } from "react-redux";
-import {
-  formInversionsControl,
-  postConversion,
-} from "@/utils/helper/inversionsFuncs";
-import { createInversion } from "@/redux/actions/inversionsActions";
+import useInvestmentForm from "@/customHooks/useInvestmentForm";
 
 export default function InvestmentFormComponent() {
   //manejan el boton de yes, no del componente
-  const [disableInput, setDisableInput] = useState(false);
-  const [selectedButton, setSelectedButton] = useState(null);
-  const dispatch = useDispatch();
-  const [cookies, setCookie] = useCookies();
 
-  const handleYesClick = () => {
-    handleButtonClick(false);
-    handleButtonClick2("Yes");
-    setDisableInput(false); // Habilitar el campo de impuestos
-  };
-
-  const handleNoClick = () => {
-    handleButtonClick(true);
-    handleButtonClick2("No");
-    setDisableInput(true); // Deshabilitar el campo de impuestos
-  };
-
-  const handleButtonClick2 = (button) => {
-    setSelectedButton(button);
-  };
-
-  const handleButtonClick = (disable) => {
-    setDisableInput(disable);
-  };
-
-  //valores del form
-  const [values, setValues] = useState({
-    name: "",
-    amount: "",
-    dayPeriod: "",
-    period: "days",
-    interest: "",
-    taxes: "",
-    earning: "", //guarda el interest en realidad, ya que no esta hecha la variable en la bd
-  });
-  //setea los valores del form a medida que cambian los componentes del form
-  const handleChange = (e) => {
-    setValues({
-      ...values,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    //llama a funcion que controla la estructura de los datos y devuelve un objeto con mensaje de error y un valor true o false en funcion del control
-    const formControl = formInversionsControl(values);
-
-    if (!formControl.booleanMessage) {
-      alert(formControl.errorMessages);
-    } else {
-      //postConversion setea los datos al formato requerido por la bd
-      if (cookies.token) {
-        const data = postConversion(values)
-        const {token} = cookies;
-        dispatch( createInversion({token, ...data}) );
-      }
-    }
-  };
+  const {disableInput,handleButtonClick,handleButtonClick2, handleChange,handleNoClick,handleSubmit,handleYesClick,selectedButton,values} = useInvestmentForm()
 
   return (
     <div>
@@ -115,6 +52,7 @@ export default function InvestmentFormComponent() {
           </div>
           <div className="column-span-1 flex flex-row">
             <button
+              type="button"
               className={`ml-2 text-black bg-black bg-opacity-10 rounded-sm h-12 w-1/2 font-extrabold ${
                 selectedButton === "Yes"
                   ? "bg-gradient-to-r from-regularPink to-boldPink hover:from-boldPink hover:to-regularPink"
@@ -125,6 +63,7 @@ export default function InvestmentFormComponent() {
               Yes
             </button>
             <button
+              type="button"
               className={`ml-2 text-black bg-black bg-opacity-10 rounded-sm h-12 w-1/2 font-extrabold ${
                 selectedButton === "No"
                   ? "bg-gradient-to-r from-regularPink to-boldPink hover:from-boldPink hover:to-regularPink"
@@ -150,11 +89,11 @@ export default function InvestmentFormComponent() {
               </span>
             </div>
           </div>
-          <input
+          <button
             className=" mt-4 text-lg bg-gradient-to-r from-regularPink to-boldPink  hover:from-boldPink hover:to-regularPink text-white rounded-sm px-5 py-2 mx-2 font-extrabold tracking-widest"
             type="submit"
             value="Submit"
-          ></input>
+          >submit</button>
         </div>
       </form>
     </div>
